refactor(domain): add featureA repository, move model to model package
