+++ app/src/main/java/com/blueprint/featureA/domain/FeatureARepository.kt
@@ -1,0 +1,8 @@
+package com.blueprint.featureA.domain
+
+import com.blueprint.featureA.domain.model.FeatureA
+
+interface FeatureARepository {
+
+    suspend fun getFeaturesA() : List<FeatureA>
+}
